+++ packages/i18n/src/index.ts
@@ -1,5 +1,6 @@
 export { en } from './en.js';
 export { fa } from './fa.js';
+export { ps } from './ps.js';
 export { format } from './format.js';
 export type { FormatOptions } from './format.js';
 export type { LocalePack, MonthNames, WeekdayNames } from './locale.js';